Few fixes for casting errors. These occurred due to commit e5f9fce01e2b26ac4ae9554e48c0317f8632499a

+++ mRemoteV1/Tools/ProcessController.cs
@@ -34,12 +34,12 @@ namespace mRemoteNG.Tools
 			if (controlHandle == IntPtr.Zero)
 				return false;
 				
-			int nCmdShow = 0;
+			uint nCmdShow = 0;
 			if (visible)
 				nCmdShow = NativeMethods.SW_SHOW;
 			else
 				nCmdShow = NativeMethods.SW_HIDE;
-            NativeMethods.ShowWindow(controlHandle, nCmdShow);
+            NativeMethods.ShowWindow(controlHandle, (int)nCmdShow);
 			return true;
 		}
